+++ rastervision_core/rastervision/core/data/utils/__init__.py
@@ -6,3 +6,4 @@ from rastervision.core.data.utils.factory import *
 from rastervision.core.data.utils.raster import *
 from rastervision.core.data.utils.rasterio import *
 from rastervision.core.data.utils.vectorization import *
+from rastervision.core.data.utils.aoi_sampler import *
