+++ rastervision_core/rastervision/core/data/utils/aoi_sampler.py
