+++ rastervision_core/requirements.txt
@@ -15,3 +15,4 @@ scikit-image==0.21.0
 boto3==1.34.14
 stackstac==0.5.0
 humanize==4.8.0
+triangle==20220202
