+++ rastervision_pytorch_learner/requirements.txt
@@ -9,7 +9,6 @@ albumentations==1.3.1
 cython==0.29.35
 pycocotools==2.0.7
 psutil==5.9.3
-triangle==20220202
 opencv-python-headless==4.9.0.80
 matplotlib==3.8.2
 tqdm==4.66.1
